Refactor utils and manual parsing
